requirements/local.txt
# -----------------------------------------
# PYTHON==3.9
# -----------------------------------------
-r base.txt

# * ---- Specific Local Development Requirements ----
psycopg2-binary==2.9.3  # https://github.com/psycopg/psycopg2
ipython==8.4.0

# * Code quality
# * ----------------------
flake8==4.0.1  # https://github.com/PyCQA/flake8
black==22.6.0  # https://github.com/psf/black
isort==5.10.1  # https://github.com/pycqa/isort
pre-commit==2.20.0 # https://github.com/pre-commit/pre-commit

# * Type check
# * ----------------------
django-stubs[compatible-mypy]==1.12.0

# * Testing
# * ----------------------
coverage~=6.4.4
factory-boy~=3.2.1
